raster/requirements.txt
flask
rasterio==1.0.18
google-cloud-storage
google-auth
